+++ Include/Intel/IndustryStandard/VirtualMemory.h
@@ -55,7 +55,7 @@ typedef union {
     UINT64    CacheDisabled        : 1;  // 0 = Cached, 1=Non-Cached
     UINT64    Accessed             : 1;  // 0 = Not accessed, 1 = Accessed (set by CPU)
     UINT64    Dirty                : 1;  // 0 = Not Dirty, 1 = written by processor on access to page
-    UINT64    PAT                  : 1;  //
+    UINT64    PAT                  : 1;  // Combines with CD, WT and MTRR to define true caching type
     UINT64    Global               : 1;  // 0 = Not global page, 1 = global page TLB not cleared on CR3 write
     UINT64    Available            : 3;  // Available for use by system software
     UINT64    PageTableBaseAddress : 40; // Page Table Base Address
@@ -80,7 +80,7 @@ typedef union {
     UINT64    MustBe1              : 1;  // Must be 1
     UINT64    Global               : 1;  // 0 = Not global page, 1 = global page TLB not cleared on CR3 write
     UINT64    Available            : 3;  // Available for use by system software
-    UINT64    PAT                  : 1;  //
+    UINT64    PAT                  : 1;  // Combines with CD, WT and MTRR to define true caching type
     UINT64    MustBeZero           : 8;  // Must be zero;
     UINT64    PageTableBaseAddress : 31; // Page Table Base Address
     UINT64    AvabilableHigh       : 11; // Available for use by system software
@@ -104,7 +104,7 @@ typedef union {
     UINT64    MustBe1              : 1;  // Must be 1
     UINT64    Global               : 1;  // 0 = Not global page, 1 = global page TLB not cleared on CR3 write
     UINT64    Available            : 3;  // Available for use by system software
-    UINT64    PAT                  : 1;  //
+    UINT64    PAT                  : 1;  // Combines with CD, WT and MTRR to define true caching type
     UINT64    MustBeZero           : 17; // Must be zero;
     UINT64    PageTableBaseAddress : 22; // Page Table Base Address
     UINT64    AvabilableHigh       : 11; // Available for use by system software
@@ -112,6 +112,19 @@ typedef union {
   } Bits;
   UINT64    Uint64;
 } PAGE_TABLE_1G_ENTRY;
+
+//
+// PAT index bits.
+//
+typedef union {
+  struct {
+    UINT8    WriteThrough  : 1;          // 0 = Write-Back caching, 1=Write-Through caching
+    UINT8    CacheDisabled : 1;          // 0 = Cached, 1=Non-Cached
+    UINT8    PAT           : 1;          // Combines with CD, WT and MTRR to define true caching type
+    UINT8    Reserved      : 5;          // Reserved
+  } Bits;
+  UINT8    Index;
+} PAT_INDEX;
 
 typedef union {
   struct {