+++ Library/OcMemoryLib/VirtualMemory.c
@@ -17,8 +17,81 @@
 
 #include <Library/UefiLib.h>
 #include <Library/BaseMemoryLib.h>
+#include <Library/MtrrLib.h>
 #include <Library/OcDebugLogLib.h>
+#include <Library/OcGuardLib.h>
 #include <Library/OcMemoryLib.h>
+
+#include <Register/Intel/Cpuid.h>
+
+//
+// Index on MTRR value.
+// (ref. table 11-8)
+//
+STATIC CHAR8  *MtrrTypeStrings[] = {
+  "UC",
+  "WC",
+  "Reserved",
+  "Reserved",
+  "WT",
+  "WP",
+  "WB"
+};
+
+//
+// Index on PAT type in PAT MSR register, which itself is
+// indexed on PAT bits from page table.
+// (ref. table 11-10)
+//
+STATIC CHAR8  *PatTypeStrings[] = {
+  "UC",
+  "WC",
+  "Reserved",
+  "Reserved",
+  "WT",
+  "WP",
+  "WB",
+  "UC-"
+};
+
+CHAR8 *
+OcGetMtrrTypeString (
+  UINT8  MtrrType
+  )
+{
+  if (MtrrType >= ARRAY_SIZE (MtrrTypeStrings)) {
+    return "Invalid";
+  } else {
+    return MtrrTypeStrings[MtrrType];
+  }
+}
+
+CHAR8 *
+OcGetPatTypeString (
+  UINT8  PatType
+  )
+{
+  if (PatType >= ARRAY_SIZE (PatTypeStrings)) {
+    return "Invalid";
+  } else {
+    return PatTypeStrings[PatType];
+  }
+}
+
+BOOLEAN
+OcIsPatSupported (
+  VOID
+  )
+{
+  CPUID_VERSION_INFO_EDX  Edx;
+
+  //
+  // Check CPUID(1).EDX[16] for PAT capability
+  //
+  AsmCpuid (CPUID_VERSION_INFO, NULL, NULL, NULL, &Edx.Uint32);
+
+  return (Edx.Bits.PAT != 0);
+}
 
 PAGE_MAP_AND_DIRECTORY_POINTER  *
 OcGetCurrentPageTable (
@@ -76,6 +149,85 @@ OcGetPhysicalAddress (
   OUT EFI_PHYSICAL_ADDRESS            *PhysicalAddr
   )
 {
+  return OcGetSetPageTableInfoForAddress (
+           PageTable,
+           VirtualAddr,
+           PhysicalAddr,
+           NULL,
+           NULL,
+           NULL,
+           FALSE
+           );
+}
+
+EFI_STATUS
+OcSetPatIndexForAddressRange (
+  IN  PAGE_MAP_AND_DIRECTORY_POINTER  *PageTable  OPTIONAL,
+  IN  EFI_VIRTUAL_ADDRESS             VirtualAddr,
+  IN  UINT64                          Length,
+  IN  PAT_INDEX                       *PatIndex
+  )
+{
+  EFI_STATUS            Status;
+  EFI_VIRTUAL_ADDRESS   EndAddr;
+  EFI_PHYSICAL_ADDRESS  PhysicalAddr;
+  UINT8                 Level;
+
+  if (OcOverflowAddU64 (VirtualAddr, Length, &EndAddr)) {
+    return EFI_INVALID_PARAMETER;
+  }
+
+  do {
+    Status = OcGetSetPageTableInfoForAddress (
+               PageTable,
+               VirtualAddr,
+               &PhysicalAddr,
+               &Level,
+               NULL,
+               PatIndex,
+               TRUE
+               );
+
+    if (EFI_ERROR (Status)) {
+      return Status;
+    }
+
+    if (VirtualAddr != PhysicalAddr) {
+      return EFI_UNSUPPORTED;
+    }
+
+    switch (Level) {
+      case 1:
+        VirtualAddr += SIZE_1GB;
+        break;
+
+      case 2:
+        VirtualAddr += SIZE_2MB;
+        break;
+
+      case 4:
+        VirtualAddr += SIZE_4KB;
+        break;
+
+      default:
+        return EFI_UNSUPPORTED;
+    }
+  } while (VirtualAddr < EndAddr);
+
+  return EFI_SUCCESS;
+}
+
+EFI_STATUS
+OcGetSetPageTableInfoForAddress (
+  IN      PAGE_MAP_AND_DIRECTORY_POINTER  *PageTable    OPTIONAL,
+  IN      EFI_VIRTUAL_ADDRESS             VirtualAddr,
+  OUT     EFI_PHYSICAL_ADDRESS            *PhysicalAddr OPTIONAL,
+  OUT     UINT8                           *Level        OPTIONAL,
+  OUT     UINT64                          *Bits         OPTIONAL,
+  IN OUT  PAT_INDEX                       *PatIndex     OPTIONAL,
+  IN      BOOLEAN                         SetPat
+  )
+{
   EFI_PHYSICAL_ADDRESS            Start;
   VIRTUAL_ADDR                    VA;
   VIRTUAL_ADDR                    VAStart;
@@ -86,6 +238,13 @@ OcGetPhysicalAddress (
   PAGE_TABLE_4K_ENTRY             *PTE4K;
   PAGE_TABLE_2M_ENTRY             *PTE2M;
   PAGE_TABLE_1G_ENTRY             *PTE1G;
+
+  //
+  // Ensure unused bits are zero.
+  //
+  if (!SetPat && (PatIndex != NULL)) {
+    PatIndex->Index = 0;
+  }
 
   if (PageTable == NULL) {
     PageTable = OcGetCurrentPageTable (NULL);
@@ -125,9 +284,33 @@ OcGetPhysicalAddress (
     //
     // 1GB PDPE
     //
-    PTE1G         = (PAGE_TABLE_1G_ENTRY *)PDPE;
-    Start         = PTE1G->Uint64 & PAGING_1G_ADDRESS_MASK_64;
-    *PhysicalAddr = Start + VA.Pg1G.PhysPgOffset;
+    PTE1G = (PAGE_TABLE_1G_ENTRY *)PDPE;
+    Start = PTE1G->Uint64 & PAGING_1G_ADDRESS_MASK_64;
+
+    if (PhysicalAddr != NULL) {
+      *PhysicalAddr = Start + VA.Pg1G.PhysPgOffset;
+    }
+
+    if (Level != NULL) {
+      *Level = 1;
+    }
+
+    if (PatIndex != NULL) {
+      if (SetPat) {
+        PTE1G->Bits.WriteThrough  = PatIndex->Bits.WriteThrough;
+        PTE1G->Bits.CacheDisabled = PatIndex->Bits.CacheDisabled;
+        PTE1G->Bits.PAT           = PatIndex->Bits.PAT;
+      } else {
+        PatIndex->Bits.WriteThrough  = (UINT8)PTE1G->Bits.WriteThrough;
+        PatIndex->Bits.CacheDisabled = (UINT8)PTE1G->Bits.CacheDisabled;
+        PatIndex->Bits.PAT           = (UINT8)PTE1G->Bits.PAT;
+      }
+    }
+
+    if (Bits != NULL) {
+      *Bits = PTE1G->Uint64;
+    }
+
     return EFI_SUCCESS;
   }
 
@@ -147,9 +330,33 @@ OcGetPhysicalAddress (
     //
     // 2MB PDE
     //
-    PTE2M         = (PAGE_TABLE_2M_ENTRY *)PDE;
-    Start         = PTE2M->Uint64 & PAGING_2M_ADDRESS_MASK_64;
-    *PhysicalAddr = Start + VA.Pg2M.PhysPgOffset;
+    PTE2M = (PAGE_TABLE_2M_ENTRY *)PDE;
+    Start = PTE2M->Uint64 & PAGING_2M_ADDRESS_MASK_64;
+
+    if (PhysicalAddr != NULL) {
+      *PhysicalAddr = Start + VA.Pg2M.PhysPgOffset;
+    }
+
+    if (Level != NULL) {
+      *Level = 2;
+    }
+
+    if (PatIndex != NULL) {
+      if (SetPat) {
+        PTE2M->Bits.WriteThrough  = PatIndex->Bits.WriteThrough;
+        PTE2M->Bits.CacheDisabled = PatIndex->Bits.CacheDisabled;
+        PTE2M->Bits.PAT           = PatIndex->Bits.PAT;
+      } else {
+        PatIndex->Bits.WriteThrough  = (UINT8)PTE2M->Bits.WriteThrough;
+        PatIndex->Bits.CacheDisabled = (UINT8)PTE2M->Bits.CacheDisabled;
+        PatIndex->Bits.PAT           = (UINT8)PTE2M->Bits.PAT;
+      }
+    }
+
+    if (Bits != NULL) {
+      *Bits = PTE2M->Uint64;
+    }
+
     return EFI_SUCCESS;
   }
 
@@ -165,10 +372,69 @@ OcGetPhysicalAddress (
     return EFI_NO_MAPPING;
   }
 
-  Start         = PTE4K->Uint64 & PAGING_4K_ADDRESS_MASK_64;
-  *PhysicalAddr = Start + VA.Pg4K.PhysPgOffset;
+  Start = PTE4K->Uint64 & PAGING_4K_ADDRESS_MASK_64;
+
+  if (PhysicalAddr != NULL) {
+    *PhysicalAddr = Start + VA.Pg4K.PhysPgOffset;
+  }
+
+  if (Level != NULL) {
+    *Level = 4;
+  }
+
+  if (PatIndex != NULL) {
+    if (SetPat) {
+      PTE4K->Bits.WriteThrough  = PatIndex->Bits.WriteThrough;
+      PTE4K->Bits.CacheDisabled = PatIndex->Bits.CacheDisabled;
+      PTE4K->Bits.PAT           = PatIndex->Bits.PAT;
+    } else {
+      PatIndex->Bits.WriteThrough  = (UINT8)PTE4K->Bits.WriteThrough;
+      PatIndex->Bits.CacheDisabled = (UINT8)PTE4K->Bits.CacheDisabled;
+      PatIndex->Bits.PAT           = (UINT8)PTE4K->Bits.PAT;
+    }
+  }
+
+  if (Bits != NULL) {
+    *Bits = PTE4K->Uint64;
+  }
 
   return EFI_SUCCESS;
+}
+
+//
+// MTRR types would ideally be passed as MTRR_MEMORY_CACHE_TYPE, but this
+// requires Library/MtrrLib.h in the .inf and .c files of everything which
+// consumes OcMemoryLib.
+//
+EFI_STATUS
+OcModifyMtrrRange (
+  IN  EFI_PHYSICAL_ADDRESS  Address,
+  IN  UINT8                 MtrrType,
+  OUT UINT64                *Length   OPTIONAL
+  )
+{
+  EFI_PHYSICAL_ADDRESS  Walker;
+  UINT8                 OriginalMtrrType;
+
+  if (Length != NULL) {
+    *Length = 0;
+  }
+
+  if (!IsMtrrSupported ()) {
+    return EFI_UNSUPPORTED;
+  }
+
+  Walker           = Address;
+  OriginalMtrrType = MtrrGetMemoryAttribute (Walker);
+  do {
+    Walker += SIZE_2MB;
+  } while (OriginalMtrrType == MtrrGetMemoryAttribute (Walker));
+
+  if (Length != NULL) {
+    *Length = Walker - Address;
+  }
+
+  return MtrrSetMemoryAttribute (Address, Walker - Address, MtrrType);
 }
 
 EFI_STATUS